fix: Ensure correct number of results per page in pagination

+++ src/main/java/com/erp/shared/dto/pagination/ResponsePageDTO.java
@@ -4,6 +4,6 @@ public record ResponsePageDTO(
         Integer number,
         Integer size,
         Integer pages,
-        Long items
+        Integer items
 ) {
 }
